Refactored instruction to use data-binding
